requirements.txt
google-analytics-data
google-genai
python-dotenv
pandas
